Game optimization
I added hit areas to the enemy now that I got the ragdoll working. Bullet to the head does triple damage now.
I made the enemy his own bullet object and script EnemyBullet. This makes more sense in terms of optimization than recycling the same game object with scripts for both the player and the enemies because of if else statements, because hundreds of bullets may fly at the same time if there are enough enemies in the scene.