Enemy AI bug fixed
Bug fixed that caused enemy to notice the sneaking player too quicly. This
happened because I had accidentally added a condition in the Scan -function of the StatePatternEnemy -script that made the enemy alert all enemies, including itself, upon seeing the player. I added a clarification to this function so that it only alerts enemies to the player's location when in CombatState